requirements.txt
defusedxml==0.7.1
exceptiongroup==1.2.2
fonttools==4.53.1
fpdf2==2.7.9
graphviz==0.20.3
iniconfig==2.0.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
pytest==8.3.2
PyYAML==6.0.1
tomli==2.0.1
